use unnamed

+++ test/e2e/app-dir/actions/app/server/actions.js
@@ -8,6 +8,6 @@ export async function dec(value) {
   return value - 1
 }
 
-export default async function double(value) {
+export default async function (value) {
   return value * 2
 }
